Vuex: refactor of vuex store
Separate origin vuex store to actions, getters and modules.
Separate menu module to a set of files.
Add a resolve path for vuex store in webpack config file.

As I set vuex store to strict mode, I fixed the related vue components.

some new states:
1. device.isTablet: for further development.
2. sidebar.hidden: if someone wants to hide siderbar for compute screen,
this will help.
3. effect: I still remember there is a style on App.vue to accelerate
page rendering in older version of vue-admin. But it will prevent affix
style. So I added a switch to fix that. If there is an affix component
fixed with postion on the page, turn effect.transalate3d to false will
let it works correctly.
