+++ client/app.js
@@ -6,6 +6,7 @@ import App from './App.vue'
 import router from './router'
 import store from './store'
 import * as filters from './filters'
+import { TOGGLE_SIDEBAR } from 'vuex-store/mutation-types'
 
 Vue.use(Resource)
 Vue.use(NProgress)
@@ -15,11 +16,10 @@ sync(store, router)
 const nprogress = new NProgress({ parent: '.nprogress-container' })
 
 const { state } = store
-const { config } = state
 
 router.beforeEach((route, redirect, next) => {
-  if (config.mobile && config.sidebar) {
-    config.sidebar = false
+  if (state.app.device.isMobile && state.app.sidebar.opened) {
+    store.commit(TOGGLE_SIDEBAR, false)
   }
   next()
 })
