--- client/store/config.js
@@ -1,4 +1,0 @@
-export default {
-  mobile: false,
-  sidebar: true
-}
