+++ client/store/index.js
@@ -1,54 +1,27 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-
 import pkg from 'package'
-import config from './config'
-import menu from '../router/menu'
+import * as actions from './actions'
+import * as getters from './getters'
+
+import app from './modules/app'
+import menu from './modules/menu'
 
 Vue.use(Vuex)
 
-const state = {
-  pkg,
-  config,
-  menu,
-  count: 0
-}
-
-const mutations = {
-  INCREMENT (state) {
-    state.count++
-  },
-  SIDEBAR (state, status = true) {
-    state.config.sidebar = status
-  }
-}
-
-const actions = {
-  INCREMENT ({ commit }) {
-    commit('INCREMENT')
-  },
-  INCREMENT_ASYNC ({ commit }) {
-    setTimeout(() => {
-      commit('INCREMENT')
-    }, 1000)
-  },
-
-  // Sidebar: show, hide, toggle
-  SHOW_SIDEBAR ({ commit }) {
-    commit('SIDEBAR', true)
-  },
-  HIDE_SIDEBAR ({ commit }) {
-    commit('SIDEBAR', false)
-  },
-  TOGGLE_SIDEBAR ({ commit, state }) {
-    commit('SIDEBAR', !state.config.sidebar)
-  }
-}
-
 const store = new Vuex.Store({
-  state,
-  mutations,
-  actions
+  strict: true,  // process.env.NODE_ENV !== 'development',
+  actions,
+  getters,
+  modules: {
+    app,
+    menu
+  },
+  state: {
+    pkg
+  },
+  mutations: {
+  }
 })
 
 export default store
