+++ client/store/modules/menu/lazyLoading.js
@@ -1,0 +1,3 @@
+// lazy loading Components
+// https://github.com/vuejs/vue-router/blob/dev/examples/lazy-loading/app.js#L8
+export const lazyLoading = (name, index = false) => () => System.import(`views/${name}${index ? '/index' : ''}.vue`)
