+++ client/store/mutation-types.js
@@ -1,0 +1,7 @@
+export const TOGGLE_DEVICE = 'TOGGLE_DEVICE'
+
+export const TOGGLE_SIDEBAR = 'TOGGLE_SIDEBAR'
+
+export const EXPAND_MENU = 'EXPAND_MENU'
+
+export const SWITCH_EFFECT = 'SWITCH_EFFECT'
